reorganised docker images, added script for MOOSE example, began adding example docs
