+++ docker_image/files/simvue.ini
@@ -1,0 +1,3 @@
+[server]
+url = https://dev02.simvue.io
+token = <your token here>
