--- docs/tutorial_advanced/files/simvue.ini
@@ -1,3 +1,0 @@
-[server]
-url = https://dev02.simvue.io
-token = <your token here>
