--- docs/tutorial_advanced/files/tutorial/step_9/moose_alerter.py
@@ -1,48 +1,0 @@
-import simvue
-import argparse
-import csv
-import time
-import os
-
-parser = argparse.ArgumentParser(description='Monitor alerts from a Simvue run.')
-parser.add_argument(
-  '--run_name', 
-  type=str,
-  help='The name of the run to monitor alerts for.'
-  )
-parser.add_argument(
-  '--time_interval', 
-  type=int,
-  help='The interval between queries to the alert status, in seconds.'
-  )
-parser.add_argument(
-  '--max_time', 
-  type=int,
-  help='The maximum time which this script will run for.'
-  )
-args = parser.parse_args()
-
-script_dir = os.path.dirname(__file__)
-
-with open(os.path.join(script_dir, 'results', 'alert_status.csv'), 'w', newline='') as csvfile: 
-    csvwriter = csv.writer(csvfile) 
-    csvwriter.writerow(['time', 'step_not_converged', 'temperature_exceeds_maximum'])
-
-time_elapsed = 0
-client = simvue.Client()
-run_id = client.get_run_id_from_name(args.run_name)
-
-while time_elapsed < args.max_time:
-    alerts = client.get_alerts(run_id)
-    with open(os.path.join(script_dir, 'results', 'alert_status.csv'), 'a', newline='') as csvfile: 
-        csvwriter = csv.writer(csvfile) 
-        csvwriter.writerow(
-            [
-                time_elapsed, 
-                ('Firing' if 'step_not_converged' in alerts else 'Normal'), 
-                ('Firing' if 'temperature_exceeds_maximum' in alerts else 'Normal')
-            ]
-          )
-    
-    time.sleep(args.time_interval)
-    time_elapsed += args.time_interval
